% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{citations}
\alias{citations}
\title{Literature citations from BETYdb}
\format{
A data frame with columns including:
\describe{
\item{id}{Citation identifier}
\item{author}{Author(s)}
\item{year}{Publication year}
\item{title}{Article/book title}
\item{journal}{Journal name}
\item{doi}{Digital Object Identifier}
}
}
\source{
\url{https://betydb.org}
}
\usage{
citations
}
\description{
Bibliographic references for data sources.
}
\keyword{datasets}
