% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{entities}
\alias{entities}
\title{Entities from BETYdb}
\format{
A data frame with entity (individual/plot) information.
}
\source{
\url{https://betydb.org}
}
\usage{
entities
}
\description{
Entities from BETYdb
}
\keyword{datasets}
