% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{managements}
\alias{managements}
\title{Management practices from BETYdb}
\format{
A data frame with management event data.
See \code{names(managements)} for all columns.
}
\source{
\url{https://betydb.org}
}
\usage{
managements
}
\description{
Management practices from BETYdb
}
\keyword{datasets}
