% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{methods}
\alias{methods}
\title{Measurement methods from BETYdb}
\format{
A data frame with method descriptions.
}
\source{
\url{https://betydb.org}
}
\usage{
methods
}
\description{
Measurement methods from BETYdb
}
\keyword{datasets}
