% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pfts}
\alias{pfts}
\title{Plant Functional Types (PFTs) from BETYdb}
\format{
A data frame. Key columns include id, name, definition.
See \code{names(pfts)} for all columns.
}
\source{
\url{https://betydb.org}
}
\usage{
pfts
}
\description{
PFT definitions used for grouping species for modeling.
}
\seealso{
\link{pfts_species}, \link{pfts_priors}
}
\keyword{datasets}
