% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pfts_priors}
\alias{pfts_priors}
\title{PFT-Prior mapping from BETYdb}
\format{
A data frame with columns:
\describe{
\item{pft_id}{Foreign key to pfts}
\item{prior_id}{Foreign key to priors}
}
}
\source{
\url{https://betydb.org}
}
\usage{
pfts_priors
}
\description{
Many-to-many relationship linking PFTs to priors.
}
\keyword{datasets}
