% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pfts_species}
\alias{pfts_species}
\title{PFT-Species mapping from BETYdb}
\format{
A data frame with columns pft_id and specie_id (plus metadata).
See \code{names(pfts_species)} for all columns.
}
\source{
\url{https://betydb.org}
}
\usage{
pfts_species
}
\description{
Many-to-many relationship linking PFTs to species.
}
\keyword{datasets}
