% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{priors}
\alias{priors}
\title{Prior distributions from BETYdb}
\format{
A data frame. Key columns include id, variable_id, distn, parama, paramb.
See \code{names(priors)} for all columns.
}
\source{
\url{https://betydb.org}
}
\usage{
priors
}
\description{
Prior probability distributions for Bayesian analysis.
}
\seealso{
\link{pfts_priors}, \link{variables}
}
\keyword{datasets}
