% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_custom.R
\name{run_composite_constraints}
\alias{run_composite_constraints}
\title{Dispatch and run all composite_constraints for a single table}
\usage{
run_composite_constraints(df, constraints)
}
\arguments{
\item{df}{data frame for the table being validated}

\item{constraints}{the composite_constraints list from the YAML for this table}
}
\value{
character vector of all error messages
}
\description{
Dispatch and run all composite_constraints for a single table
}
\keyword{internal}
