% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_custom.R
\name{run_conditional}
\alias{run_conditional}
\title{Run a conditional constraint (if field A is set, field B must also be set)}
\usage{
run_conditional(df, con)
}
\arguments{
\item{df}{data frame}

\item{con}{a single composite_constraint list element of type "conditional"}
}
\value{
character vector of error messages (empty if all rows pass)
}
\description{
Run a conditional constraint (if field A is set, field B must also be set)
}
\keyword{internal}
