% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{species}
\alias{species}
\title{Species taxonomy from BETYdb}
\format{
A data frame. Key columns include id, genus, species, scientificname,
commonname. See \code{names(species)} for all columns.
}
\source{
\url{https://betydb.org}
}
\usage{
species
}
\description{
Taxonomic information for plant species in BETYdb including USDA PLANTS
database attributes.
}
\keyword{datasets}
