% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{treatments}
\alias{treatments}
\title{Experimental treatments from BETYdb}
\format{
A data frame with treatment definitions.
See \code{names(treatments)} for all columns.
}
\source{
\url{https://betydb.org}
}
\usage{
treatments
}
\description{
Experimental treatments from BETYdb
}
\keyword{datasets}
