% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate_frictionless_layer}
\alias{validate_frictionless_layer}
\title{Run all Frictionless-layer validation across all loaded tables}
\usage{
validate_frictionless_layer(tables, datapackage_path = "datapackage.json")
}
\arguments{
\item{tables}{named list of data frames (ONLY constrained tables)}

\item{datapackage_path}{path to datapackage.json}
}
\value{
named list of character vectors; each name is a table with errors
}
\description{
Run all Frictionless-layer validation across all loaded tables
}
\keyword{internal}
