% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate_primary_key}
\alias{validate_primary_key}
\title{Validate primaryKey uniqueness for a table}
\usage{
validate_primary_key(df, schema, table_name)
}
\arguments{
\item{df}{data frame}

\item{schema}{the schema list for this resource}

\item{table_name}{string for error messages}
}
\value{
character vector of error messages
}
\description{
Validate primaryKey uniqueness for a table
}
\keyword{internal}
