% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betydata-package.R
\docType{package}
\name{betydata-package}
\alias{betydata}
\alias{betydata-package}
\title{betydata: Public Data from the BETYdb Database}
\description{
Offline access to plant trait and yield data from BETYdb.
Load datasets with \code{data(traitsview)}, \code{data(species)}, etc.
}
\seealso{
\link{traitsview}, \link{species}, \link{sites}, \link{variables}, \link{pfts}, \link{priors}
}
\author{
\strong{Maintainer}: David LeBauer \email{dlebauer@arizona.edu}

Authors:
\itemize{
  \item Akash B V \email{www.divinehome12@gmail.com}
}

Other contributors:
\itemize{
  \item University of Illinois, NCSA [copyright holder]
}

}
\keyword{internal}
