% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{citations}
\alias{citations}
\title{Literature citations from BETYdb}
\format{
A data frame. Key columns include id, author, year, title, journal, doi.
See \code{names(citations)} for all columns.
}
\source{
\url{https://betydb.org}
}
\usage{
citations
}
\description{
Bibliographic references for data sources.
}
\keyword{datasets}
