% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cultivars}
\alias{cultivars}
\title{Plant cultivars from BETYdb}
\format{
A data frame with cultivar information.
See \code{names(cultivars)} for all columns.
}
\source{
\url{https://betydb.org}
}
\usage{
cultivars
}
\description{
Plant cultivars from BETYdb
}
\keyword{datasets}
