% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cultivars_pfts}
\alias{cultivars_pfts}
\title{Cultivar-PFT mapping from BETYdb}
\format{
A data frame with columns cultivar_id and pft_id (plus metadata).
See \code{names(cultivars_pfts)} for all columns.
}
\source{
\url{https://betydb.org}
}
\usage{
cultivars_pfts
}
\description{
Many-to-many relationship linking cultivars to PFTs.
}
\keyword{datasets}
