% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{priors}
\alias{priors}
\title{Prior distributions from BETYdb}
\format{
A data frame with columns including:
\describe{
\item{id}{Prior identifier}
\item{variable_id}{Associated variable}
\item{distn}{Distribution type (e.g. "norm", "gamma")}
\item{parama, paramb, paramc}{Distribution parameters}
}
}
\source{
\url{https://betydb.org}
}
\usage{
priors
}
\description{
Prior probability distributions for Bayesian analysis.
}
\seealso{
\link{pfts_priors}, \link{variables}
}
\keyword{datasets}
