% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sites}
\alias{sites}
\title{Research sites from BETYdb}
\format{
A data frame. Key columns include id, sitename, city, state, country.
See \code{names(sites)} for all columns.
}
\source{
\url{https://betydb.org}
}
\usage{
sites
}
\description{
Geographic and metadata for research sites.
}
\keyword{datasets}
