% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{species}
\alias{species}
\title{Species taxonomy from BETYdb}
\format{
A data frame with columns including:
\describe{
\item{id}{Species identifier}
\item{spcd}{Species code}
\item{genus}{Taxonomic genus}
\item{species}{Specific epithet}
\item{scientificname}{Full scientific name}
\item{commonname}{Common name(s)}
}
}
\source{
\url{https://betydb.org}
}
\usage{
species
}
\description{
Taxonomic information for plant species in BETYdb.
}
\keyword{datasets}
