% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{variables}
\alias{variables}
\title{Variable definitions from BETYdb}
\format{
A data frame. Key columns include id, name, description, units.
See \code{names(variables)} for all columns.
}
\source{
\url{https://betydb.org}
}
\usage{
variables
}
\description{
Definitions and metadata for measured variables/traits.
}
\keyword{datasets}
